\name{plotcc}
\alias{plotcc}
\title{
Control Chart Plotting
}
\description{
Plotting a control chart.
}
\usage{
plotcc(statsII, CL, statsI = NULL)
}
\arguments{
  \item{statsII}{
the Phase II monitoring statistics.
}
  \item{CL}{
the control limit of the control chart.
}
  \item{statsI}{
(some of) the Phase I monitoring statistics.
}
}
\author{
Anh Bui
}
\references{
Bui, A. T., and Apley, D.W. (2019) \bold{spc4sts}: Statistical Process Control for Stochastic Textured Surfaces in R, working paper.
}
